syncer(dm): move shared DDL rewrite helpers
